Refactor velocity handling to use int instead of byte in Moritz.Symbols
- Updated `ConvertVoiceDefsToNoteObjects` in `Notator.cs` to use int arrays for `currentChannelVelocities` and modified logic for handling `SmallClef` objects.
- Changed parameter type of `currentVelocity` from byte to int in `GetNoteObject` methods in `StandardSymbolSet.cs` and `SymbolSet.cs`.
- Altered `VoicesPerStaff` in `PageFormat.cs` from a list of byte lists to a list of int lists.
- Updated return type of `MidiChannel` method in `SvgScore.cs` from byte to int for consistency.
- Modified `AddDynamic` and `GetDynamicString` methods in `AnchorageSymbol.cs` to use int instead of byte for velocity parameters.

+++ Moritz.Symbols/Metrics/StandardSymbolSet.cs
@@ -519,7 +519,7 @@ namespace Moritz.Symbols
         }
 
         public override NoteObject GetNoteObject(Voice voice, int absMsPosition, IUniqueDef iud, int iudIndex,
-            ref byte currentVelocity, PageFormat pageFormat)
+            ref int currentVelocity, PageFormat pageFormat)
         {
             NoteObject noteObject = null;
             CautionaryChordDef cautionaryChordDef = iud as CautionaryChordDef;
